修改 ixPoiPhoto 实体类

+++ dao/dao-glm-data/src/main/java/com/navinfo/dataservice/dao/plus/model/ixpoi/IxPoiPhoto.java
@@ -16,7 +16,7 @@ public class IxPoiPhoto extends BasicRow {
 	protected String memo ;
 	protected int tag = 1;
 	//*******2017.07.18*******
-	protected int originPid;
+	protected long originPid;
 	
 	public IxPoiPhoto (long objPid){
 		super(objPid);
@@ -72,11 +72,11 @@ public class IxPoiPhoto extends BasicRow {
 		}
 	}
 	//*******2017.07.18*******
-	public int getOriginPid() {
+	public long getOriginPid() {
 		return originPid;
 	}
 
-	public void setOriginPid(int originPid) {
+	public void setOriginPid(long originPid) {
 		this.originPid = originPid;
 	}
